dir_config() [ie no arguments] returns the APR::Table object (nc)
Hence refactor the implementation of dir_config() to use a real APR::Table
object, and not just a native hash, as the Moose hash accessor method only
provides 1 and 2 argument forms.

It's also easy to add the proper 2 argument behaviour - if the second argument
is undef, then the action is a clear [ie key deletion], not setting the value
to undef.